separate string library tests to stringtest.* files
