+++ tests/src/stringtest.h
@@ -1,0 +1,7 @@
+#ifndef __STRINGTEST_H_INCLUDED__
+#define __STRINGTEST_H_INCLUDED__
+
+void testStrings();
+void testStringCollections();
+
+#endif
